Add README.md to describe purpose
The README.md for this directory now includes details of the
various aspects of SmartSim that are demonstrated in this example.
Additionally it has conctact information for the leads on this
collaboration to contact for more information. Lastly, a brief
quick start and more detailed usage instructions that can modify
the run are included.